[ble][tcat] Add BLE 'non-advertising' state, TCAT active/standby states, and timed activation.
This adds a BLE 'non-advertising' state in which BLE is enabled for connections, but not advertising.
- If the TCAT agent is not started, advertising will be enabled by default.
- If the TCAT agent is started, advertising is under control of the TCAT agent depending on its state.

The TCAT agent, when started, can be now in standby mode or active mode. In standby, it doesn't need
to advertise and doesn't allow incoming connections. In active mode, it allows both of these.
Timed activation of the active mode is also possible. CLI commands are added to test the new modes.

Existing architecture is not changed (BLE-Secure is in control and has all APIs to TCAT agent).
It also fixes some (API) documentation and clarifies this where needed. And an error-value bugfix.

The TcatAgent is now added to instance.hpp/cpp, because that is required to use a timer.