[ble][tcat] Add BLE 'non-advertising' state and TCAT active/standby states.
This adds a BLE 'non-advertising' state in which BLE is enabled for connections, but not advertising.
- If the TCAT agent is not started, advertising will be enabled by default.
- If the TCAT agent is started, advertising is under control of BLE-secure but the TCAT agent will be
  asked through a new member function NeedsToAdvertise() if it currently needs to advertise or not.
The TCAT agent, when started, can be now in standby mode or active mode. In standby, it doesn't need
to advertise and doesn't allow incoming connections. In active both, it allows both of these.
This is a preparation for adding timer-based control and remove (TMF) activation of the TCAT agent
active/standby states.

Existing architecture is not changed (BLE-Secure is in full control and has all APIs to TCAT agent).
It also fixes some (API) documentation and clarifies this where needed. And an error-value bugfix.